mnist_ptl/requirements.txt
pytorch_lightning
torch
torchvision
metaflow-checkpoint